Add configurable request_timeout to Inngest client
We will pass this to httpx for sending events. There are other places we
set httpx timeouts that we may change later.

This allows us to configure the test to run with a shorter timeout,
reducing overall time of the test_send.py suite from ~90s to ~30s.
It still takes a while because there are up to 4 retries with
exponential backoff.