Add idempotent retries when sending events (#251)
* Start with failing test when there is an error reaching event API

Python port of Go implementation in https://github.com/inngest/inngestgo/pull/134

* Add retry loop around event sending with exponential backoff and jitter

Same exponential backoff algorithm as inngestgo

Test still fails because we're not sending the `x-inngest-event-id-seed` header yet

* Generate seed and send in `x-inngest-event-id-seed` header

Ported seed generation from inngestgo

* Rename MAX_RETRY_ATTEMPTS to MAX_SEND_ATTEMPTS

This value includes the initial send, so this is a better name

* Handle httpx.ReadTimeout in retry loop

* Add test for timing out reaching event API

Used subTest to parameterize different proxy behavior

* Parameterize tests on sync/async send events

* Add exponential backoff/retry to sync version of send too

* Factor out backoff delay computation

* Add configurable request_timeout to Inngest client

We will pass this to httpx for sending events. There are other places we
set httpx timeouts that we may change later.

This allows us to configure the test to run with a shorter timeout,
reducing overall time of the test_send.py suite from ~90s to ~30s.
It still takes a while because there are up to 4 retries with
exponential backoff.